Remove MXNet code
