+++ timeseries/src/autogluon/timeseries/models/gluonts/abstract_gluonts.py
@@ -1,4 +1,6 @@
 import logging
+import os
+import shutil
 from pathlib import Path
 from typing import Any, Callable, Dict, Iterator, List, Optional, Type
 
@@ -6,24 +8,26 @@ import gluonts
 import gluonts.core.settings
 import numpy as np
 import pandas as pd
+import torch
 from gluonts.dataset.common import Dataset as GluonTSDataset
 from gluonts.dataset.field_names import FieldName
-from gluonts.model.estimator import Estimator as GluonTSEstimator
 from gluonts.model.forecast import Forecast, QuantileForecast, SampleForecast
-from gluonts.model.predictor import Predictor as GluonTSPredictor
+from gluonts.torch.model.estimator import PyTorchLightningEstimator as GluonTSPyTorchLightningEstimator
 from gluonts.torch.model.forecast import DistributionForecast
+from gluonts.torch.model.predictor import PyTorchPredictor as GluonTSPyTorchPredictor
 from pandas.tseries.frequencies import to_offset
 
+from autogluon.common.loaders import load_pkl
 from autogluon.common.utils.log_utils import set_logger_verbosity
-from autogluon.core.hpo.constants import RAY_BACKEND
 from autogluon.core.utils import warning_filter
 from autogluon.timeseries.dataset.ts_dataframe import ITEMID, TimeSeriesDataFrame
 from autogluon.timeseries.models.abstract import AbstractTimeSeriesModel
 from autogluon.timeseries.utils.forecast import get_forecast_horizon_index_ts_dataframe
-from autogluon.timeseries.utils.warning_filters import disable_root_logger
+from autogluon.timeseries.utils.warning_filters import disable_root_logger, torch_warning_filter
 
 logger = logging.getLogger(__name__)
 gts_logger = logging.getLogger(gluonts.__name__)
+pl_loggers = [logging.getLogger(name) for name in logging.root.manager.loggerDict if "pytorch_lightning" in name]
 
 
 GLUONTS_SUPPORTED_OFFSETS = ["Y", "Q", "M", "W", "D", "B", "H", "T", "min", "S"]
@@ -127,10 +131,10 @@ class AbstractGluonTSModel(AbstractTimeSeriesModel):
         possible values.
     """
 
+    gluonts_estimator_class: Type[GluonTSPyTorchLightningEstimator]
     gluonts_model_path = "gluon_ts"
-    gluonts_estimator_class: Type[GluonTSEstimator] = None
     # datatype of floating point and integers passed internally to GluonTS
-    float_dtype: Type = np.float64
+    float_dtype: Type = np.float32
     int_dtype: Type = np.int64
     # default number of samples for prediction
     default_num_samples: int = 1000
@@ -156,7 +160,7 @@ class AbstractGluonTSModel(AbstractTimeSeriesModel):
             hyperparameters=hyperparameters,
             **kwargs,
         )
-        self.gts_predictor: Optional[GluonTSPredictor] = None
+        self.gts_predictor: Optional[GluonTSPyTorchPredictor] = None
         self.callbacks = []
         self.num_feat_static_cat = 0
         self.num_feat_static_real = 0
@@ -165,6 +169,8 @@ class AbstractGluonTSModel(AbstractTimeSeriesModel):
         self.feat_static_cat_cardinality: List[int] = []
 
     def save(self, path: str = None, verbose: bool = True) -> str:
+        # we flush callbacks instance variable if it has been set. it can keep weak references which breaks training
+        self.callbacks = []
         # The GluonTS predictor is serialized using custom logic
         predictor = self.gts_predictor
         self.gts_predictor = None
@@ -180,16 +186,12 @@ class AbstractGluonTSModel(AbstractTimeSeriesModel):
         return str(path)
 
     @classmethod
-    def load(
-        cls, path: str, reset_paths: bool = True, load_oof: bool = False, verbose: bool = True
-    ) -> "AbstractGluonTSModel":
-        model = super().load(
-            path=path,
-            reset_paths=reset_paths,
-            load_oof=load_oof,
-            verbose=verbose,
-        )
-        model.gts_predictor = GluonTSPredictor.deserialize(Path(path) / cls.gluonts_model_path)
+    def load(cls, path: str, reset_paths: bool = True, verbose: bool = True) -> "AbstractGluonTSModel":
+        with torch_warning_filter():
+            model = load_pkl.load(path=os.path.join(path, cls.model_file_name), verbose=verbose)
+            if reset_paths:
+                model.set_contexts(path)
+            model.gts_predictor = GluonTSPyTorchPredictor.deserialize(Path(path) / cls.gluonts_model_path)
         return model
 
     def _deferred_init_params_aux(self, **kwargs) -> None:
@@ -243,14 +245,41 @@ class AbstractGluonTSModel(AbstractTimeSeriesModel):
         return args
 
     def _get_estimator_init_args(self) -> Dict[str, Any]:
-        """Get GluonTS specific constructor arguments for estimator objects, an alias to
-        `self._get_model_params` for better readability."""
-        return self._get_model_params()
+        """Get GluonTS specific constructor arguments for estimator objects, an alias to `self._get_model_params`
+        for better readability."""
+        init_kwargs = super()._get_estimator_init_args()
+        # Map MXNet kwarg names to PyTorch Lightning kwarg names
+        init_kwargs.setdefault("lr", init_kwargs.get("learning_rate", 1e-3))
+        init_kwargs.setdefault("max_epochs", init_kwargs.get("epochs"))
+        return init_kwargs
 
-    def _get_estimator(self) -> GluonTSEstimator:
+    def _get_estimator(self) -> GluonTSPyTorchLightningEstimator:
         """Return the GluonTS Estimator object for the model"""
-        with warning_filter():
-            return self.gluonts_estimator_class.from_hyperparameters(**self._get_estimator_init_args())
+        # As GluonTSPyTorchLightningEstimator objects do not implement `from_hyperparameters` convenience
+        # constructors, we re-implement the logic here.
+        # we translate the "epochs" parameter to "max_epochs" for consistency in the AbstractGluonTSModel interface
+
+        init_args = self._get_estimator_init_args()
+
+        trainer_kwargs = {}
+        epochs = init_args.get("max_epochs")
+        callbacks = init_args.get("callbacks", [])
+
+        # TODO: Provide trainer_kwargs outside the function (e.g., to specify # of GPUs)?
+        if epochs is not None:
+            trainer_kwargs.update({"max_epochs": epochs})
+        trainer_kwargs.update({"callbacks": callbacks, "enable_progress_bar": False})
+        trainer_kwargs["default_root_dir"] = self.path
+
+        if torch.cuda.is_available():
+            trainer_kwargs["accelerator"] = "gpu"
+            trainer_kwargs["devices"] = 1
+
+        return from_hyperparameters(
+            self.gluonts_estimator_class,
+            trainer_kwargs=trainer_kwargs,
+            **init_args,
+        )
 
     def _to_gluonts_dataset(
         self, time_series_df: Optional[TimeSeriesDataFrame], known_covariates: Optional[TimeSeriesDataFrame] = None
@@ -308,6 +337,8 @@ class AbstractGluonTSModel(AbstractTimeSeriesModel):
         **kwargs,
     ) -> None:
         verbosity = kwargs.get("verbosity", 2)
+        for pl_logger in pl_loggers:
+            pl_logger.setLevel(logging.ERROR if verbosity <= 3 else logging.INFO)
         set_logger_verbosity(verbosity, logger=logger)
         gts_logger.setLevel(logging.ERROR if verbosity <= 3 else logging.INFO)
 
@@ -318,7 +349,6 @@ class AbstractGluonTSModel(AbstractTimeSeriesModel):
             )
 
         self._check_fit_params()
-
         # update auxiliary parameters
         self._deferred_init_params_aux(
             dataset=train_data, callbacks=self._get_callbacks(time_limit=time_limit), **kwargs
@@ -332,9 +362,14 @@ class AbstractGluonTSModel(AbstractTimeSeriesModel):
                 cache_data=True,
             )
 
+        lightning_logs_dir = Path(self.path) / "lightning_logs"
+        if lightning_logs_dir.exists() and lightning_logs_dir.is_dir():
+            logger.debug(f"Removing lightning_logs directory {lightning_logs_dir}")
+            shutil.rmtree(lightning_logs_dir)
+
     def _get_callbacks(self, time_limit: int, *args, **kwargs) -> List[Callable]:
         """Retrieve a list of callback objects for the GluonTS trainer"""
-        return []
+        return [Timer(timedelta(seconds=time_limit))] if time_limit is not None else []
 
     def predict(
         self,
@@ -396,7 +431,19 @@ class AbstractGluonTSModel(AbstractTimeSeriesModel):
     def _distribution_to_quantile_forecast(
         forecast: DistributionForecast, quantile_levels: List[float]
     ) -> QuantileForecast:
-        raise NotImplementedError
+        # Compute all quantiles in parallel instead of a for-loop
+        quantiles = torch.tensor(quantile_levels, device=forecast.distribution.mean.device).reshape(-1, 1)
+        quantile_predictions = forecast.distribution.icdf(quantiles).cpu().detach().numpy()
+        forecast_arrays = np.vstack([forecast.mean, quantile_predictions])
+        forecast_keys = ["mean"] + [str(q) for q in quantile_levels]
+
+        forecast_init_args = dict(
+            forecast_arrays=forecast_arrays,
+            start_date=forecast.start_date,
+            forecast_keys=forecast_keys,
+            item_id=str(forecast.item_id),
+        )
+        return QuantileForecast(**forecast_init_args)
 
     def _gluonts_forecasts_to_data_frame(
         self,
@@ -431,6 +478,3 @@ class AbstractGluonTSModel(AbstractTimeSeriesModel):
         result = pd.concat(result_dfs)
         result.index = forecast_index
         return TimeSeriesDataFrame(result)
-
-    def _get_hpo_backend(self):
-        return RAY_BACKEND
