+++ timeseries/tests/conftest.py
@@ -11,13 +11,8 @@ def pytest_addoption(parser):
 
 def pytest_configure(config):
     config.addinivalue_line("markers", "slow: mark test as slow to run")
-    # known mxnet warnings
-    config.addinivalue_line("filterwarnings", "ignore:In accordance with NEP 32:DeprecationWarning")
-    config.addinivalue_line("filterwarnings", "ignore:.np.bool:DeprecationWarning")
 
     # known gluonts warnings
-    config.addinivalue_line("filterwarnings", "ignore::DeprecationWarning:.*gluonts.mx.distribution.*")
-    config.addinivalue_line("filterwarnings", "ignore::DeprecationWarning:.*gluonts.mx.trainer.*")
     config.addinivalue_line("filterwarnings", "ignore:Using `json`-module:UserWarning")
 
     # pandas future warnings on timestamp freq being deprecated
